backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
pydantic>=2.8.0
pydantic-settings>=2.4.0
requests>=2.32.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
sqlalchemy>=2.0
psycopg2-binary>=2.9
python-jose[cryptography]>=3.3
